consumer_2/requirements.txt
kafka-python
uvicorn
fastapi
requests